+++ CHANGELOG.md
@@ -15,6 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SPFarm
   - Updated to run cmdlet `Update-SPFlightsConfigFile` on SharePoint Subscription.
 
+### Fixed
+
+- SPCertificateSettings
+  - Fixed an error where the command failed to add
+    SPCertificateNotificationContacts when there are currently none set.
+
 ## [5.4.0] - 2023-04-04
 
 ### Fixed